services/rag-indexation/metadata-pipelines/pyproject.toml
[project]
name = "metadata-pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyodbc>=5.0.1",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
# Añade el directorio raíz del proyecto al PYTHONPATH
package = false

[tool.uv.sources]

[tool.setuptools]
py-modules = []

# Ejecutar con: uv run python -m src.main
